Enum vlive::Error

pub enum Error {
    Hyper(HyperError),
    Uri(UriError),
    Reqwest(ReqwestError),
    Json(JsonError),
    Io(IoError),
    Vlive(String),
}

Common error type used throughout the library, to be used as a holder for errors from various other libraries.

Variants
Hyper(HyperError)

A hyper crate error.

Uri(UriError)

An error from hyper while parsing a URI.

Reqwest(ReqwestError)

A reqwest crate error.

Json(JsonError)

A serde_json crate error.

Io(IoError)

A std::io module error.

Vlive(String)

A vlive crate error.