Struct vlive::model::channel::ChannelInfo

pub struct ChannelInfo {
    pub channel_seq: u32,
    pub basic_channel_seq: u32,
    pub channel_code: String,
    pub channel_plus_type: ChannelType,
    pub channel_name: String,
    pub representative_color: String,
    pub channel_profile_image: String,
    pub background_color: String,
    pub channel_cover_image: String,
    pub fan_count: u32,
    pub comment: String,
    pub prohibited_word_like: String,
    pub prohibited_word_exact: String,
    pub sns_share_img: String,
    pub banner_show_yn: bool,
    pub qrcode: String,
    pub upcoming_show_yn: bool,
}

Information on a channel

Fields
channel_seq: u32

Channel ID, used in API queries.

basic_channel_seq: u32

Channel ID of basic channel if this is a CHANNEL+ channel.

channel_code: String

Channel code, used in URLs.

channel_plus_type: ChannelType

Type of channel.

channel_name: String

Name of channel.

representative_color: String

Color of channel.

channel_profile_image: String

Channel profile image

background_color: String

Channel background color

channel_cover_image: String

Channel color image

fan_count: u32

Number of followers

comment: String

Description of page, usually just "Welcome to {name} Channel!"

prohibited_word_like: String

Usually empty

prohibited_word_exact: String

Usually empty

sns_share_img: String

Image used for sharing

banner_show_yn: bool qrcode: String

QR code that gives a link to this channel

upcoming_show_yn: bool

If there are upcoming videos

Methods

impl ChannelInfo

pub fn url(&self) -> String

Gets the URL to this video.