Struct vlive::model::channel::VideoListItem

pub struct VideoListItem {
    pub video_seq: u32,
    pub video_type: String,
    pub title: String,
    pub play_count: u32,
    pub like_count: u32,
    pub comment_count: u32,
    pub thumbnail: String,
    pub pick_sort_order: u32,
    pub screen_orientation: String,
    pub will_start_at: DateTime<FixedOffset>,
    pub will_end_at: DateTime<FixedOffset>,
    pub created_at: Option<DateTime<FixedOffset>>,
    pub upcoming_yn: String,
    pub special_live_yn: bool,
    pub live_thumb_yn: bool,
    pub product_id: String,
    pub package_product_id: String,
    pub product_type: String,
    pub play_time: u32,
    pub channel_plus_public_yn: bool,
    pub expose_status: String,
    pub on_air_start_at: DateTime<FixedOffset>,
}

Information on a video

Fields
video_seq: u32

The ID of the video, used in the URL.

video_type: String

Type of video, either VOD or LIVE

title: String

Title of the video.

play_count: u32

Number of plays this video has.

like_count: u32

Number of likes or hearts this video has.

comment_count: u32

Number of comments this video has.

thumbnail: String

The thumbnail URL for this video.

pick_sort_order: u32 screen_orientation: String

Screen orientation of this video, either HORIZONTAL or VERTICAL

will_start_at: DateTime<FixedOffset>

Upload date of this video(?) This video may not have been visible at this point.

will_end_at: DateTime<FixedOffset>

End time of this video, usually sometime in 2099 so this isn't really useful.

created_at: Option<DateTime<FixedOffset>> upcoming_yn: String special_live_yn: bool

If this is a "special" live video.

live_thumb_yn: bool

If this video has a live thumbnail.

product_id: String

VLIVE+ product ID if this is a VLIVE+ video, otherwise it's empty.

package_product_id: String

VLIVE+ package ID if this is a VLIVE+ video, otherwise it's empty.

product_type: String

If this is a VLIVE+ product video.

play_time: u32

Duration of the video in seconds.

channel_plus_public_yn: bool

If this is a basic or CHANNEL+ video

expose_status: String on_air_start_at: DateTime<FixedOffset>

Date when this video was available.

Methods

impl VideoListItem

pub fn url(&self) -> String

Gets the URL to this video.

pub fn is_live(&self) -> bool

Checks if this video is currently live.